Add files via upload
I just completed TIC TAC TOE project. Tic-Tac-Toe is a classic two-player game that involves a 3x3 grid. Players take turns marking a cell in the grid with their symbol, either 'X' or 'O'. The player who first succeeds in placing three of their marks in a horizontal, vertical, or diagonal row wins the game. If all nine squares are filled without either player achieving three in a row, the game is considered a tie.